+++ config/api/auth/request-resource-delete/fields-post.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'force' => [
+        'field' => 'force',
+        'title' => 'auth/request-resource-delete/fields.title-force',
+        'description' => 'auth/request-resource-delete/fields.description-force',
+        'type' => 'boolean',
+        'required' => true
+    ]
+];