+++ resources/lang/en/auth/request-resource-delete/fields.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'title-force' => 'Force',
+    'description-force' => 'Force the deletion, it is required that this be set to true if multiple users have access to the parent resource-type',
+];